% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_probs.R
\name{compute_mixing_weights}
\alias{compute_mixing_weights}
\title{Compute mixing probabilities for a partition}
\usage{
compute_mixing_weights(
  y,
  covariates,
  age,
  tin,
  combins,
  estimates,
  L = 8,
  init = c(1, 0, 0, 0, 0, 0, 0, 0),
  threads = 8,
  cpp_file = NULL,
  save_path = NULL
)
}
\arguments{
\item{y}{Observation matrix (individuals x time).}

\item{covariates}{Covariate matrix (individuals x 11).}

\item{age}{Age indicator array (individuals x 2 x time). If a 3-level array
is supplied, the third level is used; if a 2-level array is supplied, the
third level is treated as all zeros.}

\item{tin}{Time-in indicator array (individuals x 2 x time).}

\item{combins}{Register combination matrix.}

\item{estimates}{Parameter estimates matrix (bootstraps x parameters).}

\item{L}{Number of registers.}

\item{init}{Initial state distribution.}

\item{threads}{Number of threads for RcppParallel.}

\item{cpp_file}{Optional path to a C++ file to source via Rcpp::sourceCpp.}

\item{save_path}{Optional path to save the mixing weights as an RDS file.}
}
\value{
A 3D array of mixing weights (individuals x 2 x bootstraps).
}
\description{
Computes the mixing probabilities for each individual using the compiled
C++ routine \code{mixing_probs}. This is packaged as a wrapper function that
accepts inputs directly and optionally saves the results.
}
