% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_out.R
\name{create_list_out}
\alias{create_list_out}
\title{Creation of list of people that deregister from the population.}
\usage{
create_list_out(pop, presences, departures, formula_prob, coef_values)
}
\arguments{
\item{pop}{A baseline population from which one will select individuals to
arrive and leave in a number of years.}

\item{presences}{A matrix of presences of all the individuals in the
population for a number of years.}

\item{departures}{A vector of departure times of all the individuals in the
population for a number of years.}

\item{formula_prob}{A linear predictor formula to calculate the probability
of deregistering.}

\item{coef_values}{The values for each coefficient to be used in the linear
predictor defined in `formula_prob`. The coefficients will be used
considering a logit link function.}
}
\value{
A matrix with the information of the deregistering (1) or non
  non deregistering (0) for each individual of the population considering
  a number of years of observation.
}
\description{
Given a population and a list of presences, `create_list_out()` creates a
  list of individuals who deregister.
}
